server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adk-backend"
version = "1.0.0"
description = "ADK Document Processing Backend API"
authors = [{name = "User", email = "user@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.2",
    "google-adk-agents>=0.1.0",
    "docling",
    "pydantic>=2.5.0",
    "streamlit>=1.28.1",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0"
]

[project.scripts]
start-api = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
test-ui = "streamlit run test_streamlit.py --server.port 8501"

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "usecases*", "routes*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
